Move ParseVisibility to query package
